scene-text-telescope/requirement.txt
easydict==1.9
editdistance==0.5.3
lmdb==1.2.1
matplotlib==3.3.4
numpy
opencv-python==4.5.2.52
Pillow==8.2.0
six
tensorboard==2.5.0
tensorboard-data-server==0.6.1
tensorboard-plugin-wit==1.8.0
torch==1.2.0
torchvision==0.2.1
tqdm==4.61.0
pyyaml
ipython
future
protobuf<=3.20.0